fetch README and update for new build
